a bit of a big commits, removed all require rubygems from code, added the bitly module to the main bitly.rb, moved v2 files to there own dir, and some minor minor cleanup
