+++ lib/bitly/v2.rb
@@ -1,0 +1,3 @@
+require 'bitly/v2/utils'
+require 'bitly/v2/client'
+require 'bitly/v2/url'
