+++ lib/bitly/v3.rb
@@ -1,9 +1,6 @@
-$:.unshift File.dirname(__FILE__)
-
 require 'httparty'
 require 'cgi'
 
-require 'v3/bitly'
-require 'v3/client'
-require 'v3/url'
-require 'v3/missing_url'
+require 'bitly/v3/client'
+require 'bitly/v3/url'
+require 'bitly/v3/missing_url'
